services/worker/pyproject.toml
[tool.poetry]
name = "ingestion-worker"
version = "0.1.0"
description = "Asynchronous worker for article ingestion and NLP analysis with comprehensive type safety."
authors = ["Your Name <you@example.com>"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
pika = "^1.3.0"
python-dotenv = "^1.0.0"
langchain = "^0.3.0"
langchain-community = "^0.3.0"
langchain-openai = "^0.2.0"
yake = "^0.4.8"
vaderSentiment = "^3.3.2"
chromadb = "^0.5.0"
pydantic = "^2.7.1"
pydantic-settings = "^2.2.1"
redis = "^5.0.0"
beautifulsoup4 = "^4.12.0"
lxml = "^5.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.0.0"
isort = "^5.12.0"
mypy = "^1.5.0"
pylint = "^2.17.0"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"